langchain/requirements.txt
langchain_openai
langchain
streamlit
openai
python-dotenv
langchain_community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
arxiv
wikipedia
langchainhub
groq
langchain_groq